After Bug fixed:  Ping OK2~ Enhance PTP message handling and logging
- Introduced new PTP message type constants for better classification.
- Refactored `dm9051_ptp_one_step` to improve clarity and functionality, including detailed logging for PTP packet processing.
- Updated `dm9051_hwtstamp_to_skb` to handle PTP message types more robustly and added packet dumping capabilities for debugging.

+++ lnx_dm9051_kt6_r2503_v3.9.5e_ptp/1_v3.9.5b/dm9051.c
@@ -2134,7 +2134,7 @@ static int dm9051_loop_tx(struct board_info *db)
 			/* ptpc */
 			#if 1 //0
 			#ifdef DMPLUG_PTP
-			db->ptp_mode = dm9051_ptp_one_step(skb); //_15888_,
+			db->ptp_mode = (int) dm9051_ptp_one_step(skb); //_15888_,
 			db->tcr_wr = dm9051_tcr_wr(skb, db); //_15888_,
 			#endif
 			#endif
